--- common/constant/generic_serialization.go
@@ -1,6 +1,0 @@
-package constant
-
-const (
-	GENERIC_SERIALIZATION_DEFAULT  = "true"
-	GENERIC_SERIALIZATION_PROTOBUF = "protobuf-json"
-)
