+++ common/constant/key.go
@@ -356,3 +356,9 @@ const (
 	// SERVICE_DISCOVERY_KEY indicate which service discovery instance will be used
 	SERVICE_DISCOVERY_KEY = "service_discovery"
 )
+
+// Generic Filter
+const (
+	GenericSerializationDefault  = "true"
+	GenericSerializationProtobuf = "protobuf-json"
+)
