+++ common/constant/serialization.go
