+++ filter/generic/filter_test.go
@@ -20,9 +20,7 @@ package generic
 import (
 	"context"
 	"net/url"
-	"reflect"
 	"testing"
-	"time"
 )
 
 import (
@@ -42,7 +40,7 @@ import (
 func TestInvoke(t *testing.T) {
 	invokeUrl := common.NewURLWithOptions(
 		common.WithParams(url.Values{}),
-		common.WithParamsValue(constant.GENERIC_KEY, constant.GENERIC_SERIALIZATION_DEFAULT))
+		common.WithParamsValue(constant.GENERIC_KEY, constant.GenericSerializationDefault))
 	filter := &Filter{}
 
 	ctrl := gomock.NewController(t)
@@ -59,7 +57,7 @@ func TestInvoke(t *testing.T) {
 			assert.Equal(t, "hello", args[0])
 			assert.Equal(t, "java.lang.String", args[1].([]interface{})[0].(string))
 			assert.Equal(t, "arg1", args[2].([]interface{})[0].(string))
-			assert.Equal(t, constant.GENERIC_SERIALIZATION_DEFAULT, invocation.AttachmentsByKey(constant.GENERIC_KEY, ""))
+			assert.Equal(t, constant.GenericSerializationDefault, invocation.AttachmentsByKey(constant.GENERIC_KEY, ""))
 			return &protocol.RPCResult{}
 		})
 
@@ -71,7 +69,7 @@ func TestInvoke(t *testing.T) {
 func TestInvokeWithGenericCall(t *testing.T) {
 	invokeUrl := common.NewURLWithOptions(
 		common.WithParams(url.Values{}),
-		common.WithParamsValue(constant.GENERIC_KEY, constant.GENERIC_SERIALIZATION_DEFAULT))
+		common.WithParamsValue(constant.GENERIC_KEY, constant.GenericSerializationDefault))
 	filter := &Filter{}
 
 	ctrl := gomock.NewController(t)
@@ -92,119 +90,10 @@ func TestInvokeWithGenericCall(t *testing.T) {
 			assert.Equal(t, "hello", args[0])
 			assert.Equal(t, "java.lang.String", args[1].([]string)[0])
 			assert.Equal(t, "arg1", args[2].([]string)[0])
-			assert.Equal(t, constant.GENERIC_SERIALIZATION_DEFAULT, invocation.AttachmentsByKey(constant.GENERIC_KEY, ""))
+			assert.Equal(t, constant.GenericSerializationDefault, invocation.AttachmentsByKey(constant.GENERIC_KEY, ""))
 			return &protocol.RPCResult{}
 		})
 
 	result := filter.Invoke(context.Background(), mockInvoker, genericInvocation)
 	assert.NotNil(t, result)
-}
-
-func TestObjToMap(t *testing.T) {
-	var testData struct {
-		AaAa string `m:"aaAa"`
-		BaBa string
-		CaCa struct {
-			AaAa string
-			BaBa string `m:"baBa"`
-			XxYy struct {
-				xxXx string `m:"xxXx"`
-				Xx   string `m:"xx"`
-			} `m:"xxYy"`
-		} `m:"caCa"`
-		DaDa time.Time
-		EeEe int
-	}
-	testData.AaAa = "1"
-	testData.BaBa = "1"
-	testData.CaCa.BaBa = "2"
-	testData.CaCa.AaAa = "2"
-	testData.CaCa.XxYy.xxXx = "3"
-	testData.CaCa.XxYy.Xx = "3"
-	testData.DaDa = time.Date(2020, 10, 29, 2, 34, 0, 0, time.Local)
-	testData.EeEe = 100
-	m := objToMap(testData).(map[string]interface{})
-	assert.Equal(t, "1", m["aaAa"].(string))
-	assert.Equal(t, "1", m["baBa"].(string))
-	assert.Equal(t, "2", m["caCa"].(map[string]interface{})["aaAa"].(string))
-	assert.Equal(t, "3", m["caCa"].(map[string]interface{})["xxYy"].(map[string]interface{})["xx"].(string))
-
-	assert.Equal(t, reflect.Map, reflect.TypeOf(m["caCa"]).Kind())
-	assert.Equal(t, reflect.Map, reflect.TypeOf(m["caCa"].(map[string]interface{})["xxYy"]).Kind())
-	assert.Equal(t, "2020-10-29 02:34:00", m["daDa"].(time.Time).Format("2006-01-02 15:04:05"))
-	assert.Equal(t, 100, m["eeEe"].(int))
-}
-
-type testStruct struct {
-	AaAa string
-	BaBa string `m:"baBa"`
-	XxYy struct {
-		xxXx string `m:"xxXx"`
-		Xx   string `m:"xx"`
-	} `m:"xxYy"`
-}
-
-func TestObjToMap_Slice(t *testing.T) {
-	var testData struct {
-		AaAa string `m:"aaAa"`
-		BaBa string
-		CaCa []testStruct `m:"caCa"`
-	}
-	testData.AaAa = "1"
-	testData.BaBa = "1"
-	var tmp testStruct
-	tmp.BaBa = "2"
-	tmp.AaAa = "2"
-	tmp.XxYy.xxXx = "3"
-	tmp.XxYy.Xx = "3"
-	testData.CaCa = append(testData.CaCa, tmp)
-	m := objToMap(testData).(map[string]interface{})
-
-	assert.Equal(t, "1", m["aaAa"].(string))
-	assert.Equal(t, "1", m["baBa"].(string))
-	assert.Equal(t, "2", m["caCa"].([]interface{})[0].(map[string]interface{})["aaAa"].(string))
-	assert.Equal(t, "3", m["caCa"].([]interface{})[0].(map[string]interface{})["xxYy"].(map[string]interface{})["xx"].(string))
-
-	assert.Equal(t, reflect.Slice, reflect.TypeOf(m["caCa"]).Kind())
-	assert.Equal(t, reflect.Map, reflect.TypeOf(m["caCa"].([]interface{})[0].(map[string]interface{})["xxYy"]).Kind())
-}
-
-func TestObjToMap_Map(t *testing.T) {
-	var testData struct {
-		AaAa   string
-		Baba   map[string]interface{}
-		CaCa   map[string]string
-		DdDd   map[string]interface{}
-		IntMap map[int]interface{}
-	}
-	testData.AaAa = "aaaa"
-	testData.Baba = make(map[string]interface{})
-	testData.CaCa = make(map[string]string)
-	testData.DdDd = nil
-	testData.IntMap = make(map[int]interface{})
-
-	testData.Baba["kk"] = 1
-	var structData struct {
-		Str string
-	}
-	structData.Str = "str"
-	testData.Baba["struct"] = structData
-	testData.Baba["nil"] = nil
-	testData.CaCa["k1"] = "v1"
-	testData.CaCa["kv2"] = "v2"
-	testData.IntMap[1] = 1
-	m := objToMap(testData)
-
-	assert.Equal(t, reflect.Map, reflect.TypeOf(m).Kind())
-	mappedStruct := m.(map[string]interface{})
-	assert.Equal(t, reflect.String, reflect.TypeOf(mappedStruct["aaAa"]).Kind())
-	assert.Equal(t, reflect.Map, reflect.TypeOf(mappedStruct["baba"]).Kind())
-	assert.Equal(t, reflect.Map, reflect.TypeOf(mappedStruct["baba"].(map[interface{}]interface{})["struct"]).Kind())
-	assert.Equal(t, "str", mappedStruct["baba"].(map[interface{}]interface{})["struct"].(map[string]interface{})["str"])
-	assert.Equal(t, nil, mappedStruct["baba"].(map[interface{}]interface{})["nil"])
-	assert.Equal(t, reflect.Map, reflect.TypeOf(mappedStruct["caCa"]).Kind())
-	assert.Equal(t, reflect.Map, reflect.TypeOf(mappedStruct["ddDd"]).Kind())
-	intMap := mappedStruct["intMap"]
-	assert.Equal(t, reflect.Map, reflect.TypeOf(intMap).Kind())
-	assert.Equal(t, 1, intMap.(map[interface{}]interface{})[1])
 }
