+++ filter/generic/util.go
@@ -1,0 +1,31 @@
+package generic
+
+import (
+	"strings"
+)
+
+import (
+	"dubbo.apache.org/dubbo-go/v3/common/constant"
+	"dubbo.apache.org/dubbo-go/v3/protocol"
+)
+
+// isCallingToGenericService check if it calls to a generic service
+func isCallingToGenericService(invoker protocol.Invoker, invocation protocol.Invocation) bool {
+	return isGeneric(invoker.GetURL().GetParam(constant.GENERIC_KEY, "")) &&
+		invocation.MethodName() != constant.GENERIC
+}
+
+// isMakingAGenericCall check if it is making a generic call to a generic service
+func isMakingAGenericCall(invoker protocol.Invoker, invocation protocol.Invocation) bool {
+	return isGeneric(invoker.GetURL().GetParam(constant.GENERIC_KEY, "")) &&
+		invocation.MethodName() == constant.GENERIC &&
+		invocation.Arguments() != nil &&
+		len(invocation.Arguments()) == 3
+}
+
+// isGeneric receives a generic field from url of invoker to determine whether the service is generic or not
+func isGeneric(generic string) bool {
+	lowerGeneric := strings.ToLower(generic)
+	return lowerGeneric == constant.GenericSerializationDefault ||
+		lowerGeneric == constant.GenericSerializationProtobuf
+}
